+++ src/app/components/PageContent.tsx
@@ -10,14 +10,7 @@ import { useState } from "react";
 import { Button } from "src/components/Button";
 import { GroupsApis } from "src/lib/apis/groupsApis";
 import { GroupCategory } from "../types";
-
-const CATEGORIES: { key: GroupCategory; text: string }[] = [
-  { key: "ALL", text: "전체" },
-  { key: "SMARTPHONE", text: "스마트폰" },
-  { key: "PPT", text: "파워포인트" },
-  { key: "VIDEO", text: "영상편집" },
-  { key: "PS", text: "포토샵" },
-];
+import { CATEGORIES } from "@app/constants";
 
 interface Group {
   id: number;
