+++ src/app/constants/index.ts
@@ -1,0 +1,9 @@
+import { GroupCategory } from "@app/types";
+
+export const CATEGORIES: { key: GroupCategory; text: string }[] = [
+  { key: "ALL", text: "전체" },
+  { key: "SMARTPHONE", text: "스마트폰" },
+  { key: "PPT", text: "파워포인트" },
+  { key: "VIDEO", text: "영상편집" },
+  { key: "PS", text: "포토샵" },
+];
