+++ src/components/hooks/useAuthKakao.ts
@@ -1,11 +1,11 @@
 import { isLoggedInAtom, userAtom } from "@app/GlobalProvider";
-import { setRefreshTokenToCookie } from "@components/serverAuth";
+import { setRefreshTokenToCookie } from "@components/server/serverAuth";
 import { useMutation } from "@tanstack/react-query";
 import axios from "axios";
 import { useSetAtom } from "jotai";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useEffect } from "react";
-import { KakaoAuth } from "src/lib/apis/KakaoAuth";
+import { KakaoAuth } from "src/lib/apis/kakao/KakaoAuth";
 import { AuthUserApis } from "src/lib/apis/authUserApis";
 
 const useAuthKakao = () => {
