+++ src/components/modal/CreatePostModal/types/index.ts
