+++ src/components/modal/LevelTestModal/constants/index.ts
@@ -1,0 +1,16 @@
+export const DEVICE_ITEMS = {
+  SMARTPHONE: { value: "스마트폰", isSelected: false },
+  LABTOP: { value: "노트북", isSelected: false },
+  TABLET: { value: "태블릿", isSelected: false },
+};
+export const SW_ITEMS = {
+  KAKAO: { value: "카카오톡", isSelected: false },
+  YOUTUBE: { value: "유튜브", isSelected: false },
+  BAND: { value: "밴드", isSelected: false },
+};
+
+export const LEVEL = [
+  { value: "초급", text: "디지털 기기와 아직 익숙하지 않은 상태예요!" },
+  { value: "증급", text: "디지털 기기와 어느정도 익숙한 수준이에요!" },
+  { value: "고급", text: "디지털 기기와 매우 친숙한 수준이에요!" },
+];
