+++ src/components/modal/LevelTestModal/index.tsx
@@ -6,22 +6,15 @@ import LevelSVG from "public/assets/illusts/levelTest.svg";
 import { useMemo } from "react";
 import { Button } from "src/components/Button";
 import Modal from "src/components/Modal";
+import { DEVICE_ITEMS, LEVEL, SW_ITEMS } from "./constants";
 
 interface Items {
   [key: string]: { value: string; isSelected: boolean };
 }
 const modalOpenAtom = atom<boolean>(false);
 const levelAtom = atomWithReset<number>(0);
-const deviceItemsAtom = atomWithReset<Items>({
-  SMARTPHONE: { value: "스마트폰", isSelected: false },
-  LABTOP: { value: "노트북", isSelected: false },
-  TABLET: { value: "태블릿", isSelected: false },
-});
-const swItemsAtom = atomWithReset<Items>({
-  KAKAO: { value: "카카오톡", isSelected: false },
-  YOUTUBE: { value: "유튜브", isSelected: false },
-  BAND: { value: "밴드", isSelected: false },
-});
+const deviceItemsAtom = atomWithReset<Items>(DEVICE_ITEMS);
+const swItemsAtom = atomWithReset<Items>(SW_ITEMS);
 
 const LevelTestModal = () => {
   const [openAtom, setOpenAtom] = useAtom(modalOpenAtom);
@@ -154,12 +147,6 @@ LevelTestModal.TestCotent = LevelTestContent;
 
 const LevelResultContent = () => {
   const [level, setLevel] = useAtom(levelAtom);
-
-  const LEVEL = [
-    { value: "초급", text: "디지털 기기와 아직 익숙하지 않은 상태예요!" },
-    { value: "증급", text: "디지털 기기와 어느정도 익숙한 수준이에요!" },
-    { value: "고급", text: "디지털 기기와 매우 친숙한 수준이에요!" },
-  ];
 
   return (
     <Modal.Content className="w-[550px] h-[70%] flex flex-col py-10 px-8">
